Split tests in different classes
